chore: Update SQLNode struct to include quotation field
This commit updates the `SqlNode` struct in multiple files to include a new field called `quotation`. The `quotation` field is used to store the quotation type used in the SQL content. It can be either a single quotation mark (`"`) or a raw string (`r#""`). This change allows for more accurate parsing and handling of SQL statements with different quotation types.