refactor: Add method_line field to SqlNode struct
This commit adds a new field called `method_line` to the `SqlNode` struct in the `config` and `sql-extraction` modules. The `method_line` field represents the line number where the SQL code is located within the method or function. This information is useful for debugging and analyzing SQL code. By including the `method_line` field, we improve the accuracy and context of the SQL extraction process.