refactor: Update formatSql.ts to use "tabSize" for SQL formatting
This commit updates the formatSql.ts file to use the "tabSize" property for SQL formatting instead of the deprecated "indent" property. It also adds a default tab size of 4 if the "tabSize" property is not specified in the configuration. This change improves the consistency and clarity of the SQL formatting options.