chore: Update npm dependencies to jest latest versions
This commit updates the npm dependencies in the package.json and pnpm-lock.yaml files to their latest stable versions. It includes updates for @jest/types, @types/jest, jest, jest-environment-node, and jest-mock-vscode. Updating the dependencies ensures that the project is using the most up-to-date versions and can take advantage of bug fixes and new features.

Co-authored-by: dependabot[bot] <support@github.com>